Reorganize, update bioreactor UI for demo.
