+++ imports/things/BaseThing/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "cg-basething",
+  "version": "0.1.0",
+  "description": "A base thing class for making UI components with Grow-IoT",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Common Garden",
+  "license": "MIT"
+}
