+++ imports/things/Camera/ImageComponent.jsx
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { Meteor } from 'meteor/meteor';
-import ImageUpload from '../app/components/images/ImageUpload';
-import ImageGrid from '../app/components/images/ImageGrid';
-import ImageOne from '../app/components/images/ImageOne';
+import ImageUpload from '../../app/components/images/ImageUpload';
+import ImageGrid from '../../app/components/images/ImageGrid';
+import ImageOne from '../../app/components/images/ImageOne';
 import {Card, CardActions, CardHeader, CardMedia, CardTitle, CardText} from 'material-ui/Card';
 
 // TODO: implement show more...
