+++ imports/things/index.js
@@ -1,13 +1,14 @@
 import GrowHub from './GrowHub/GrowHub';
 import TestDevice from './Misc/TestDevice.jsx';
 import SmartLight from './SmartLight/SmartLight.jsx';
-import ImageComponent from './ImageComponent';
+import ImageComponent from './Camera/ImageComponent';
 import TentacleExample from './Misc/TentacleExample';
 import CameraComponent from './Camera/CameraComponent';
 import DrDose from './Misc/DrDose';
 import ClimateSensor from './Misc/ClimateSensor';
 import PlusFarm from './Misc/PlusFarm';
 import Thermostat from './Misc/Thermostat';
+import BioReactor from './BioReactor/BioReactor';
 
 const components = {
 	GrowHub,
